CV
hef has a string S with length N. He needs to find the number of indices i (1≤i≤N−1) such that the i-th character of this string is a consonant and the i+1-th character is a vowel. However, he is busy, so he asks for your help.

Note: The letters 'a', 'e', 'i', 'o', 'u' are vowels; all other lowercase English letters are consonants.