UEFA Champions League
The UEFA Champions League is the most prestigious annual sports competition in the world. In the group stage of this competition, European football clubs are divided into 8 groups; there are four teams in each group. The teams in each group are ranked based on the matches they play against each other, according to the following rules:

Based on the results of matches, teams are awarded points. Each football match is played between a home team and an away team. If one of the teams scores more goals than the other, this team gains 3 points and the other team gains 0 points. In case of a tie (if both teams score the same number of goals), each of those teams gains 1 point.
The goal difference of a team is the number of goals it scored minus the number of goals it received, regardless if it scored/received them as the home team or as the away team.
Between any two teams, the team with more points is ranked higher.
If they have the same number of points (in case of a tie), the team with higher goal difference is ranked higher.
Each team plays two matches against every other team in its group ― one match as the home team and one match as the away team. You are given the number of goals scored by each team for all twelve matches in one group. Determine the leaders of this group ― the first and second top-ranked team. It is guaranteed that there are no ties for either of these places (for the given results of the matches).