Video Game
You are playing a video game in which several stacks of boxes are lined up on the floor, with a crane on top to rearrange the boxes.Move one position left (does nothing if already at the leftmost position)


• Move one position right (does nothing if already at the rightmost position)


• Pick up a box from the current stack (does nothing if the crane already has a box)


• Drop a box on the current stack (does nothing if the crane doesn't already have a box)


Further, there is a limit H on the number of boxes on each stack. If a 'drop' command would result in a stack having more than H boxes, the crane ignores this drop command. If the current stack has no boxes, a 'pick up' command is ignored.


You are given the initial number of boxes in each stack and the sequence of operations performed by the crane. You have to compute the final number of boxes in each stack.