Sherlock and Squares
Watson likes to challenge Sherlock's math ability. He will provide a starting and ending value describing a range of integers. Sherlock must determine the number of square integers within that range, inclusive of the endpoints.

Note: A square integer is an integer which is the square of an integer, e.g. .1,4,9,16,25

For example, the range is a=24 and b=49 inclusive. There are three square integers in the range:25,36  and 49.