Flatland Space Stations
latland is a country with a number of cities, some of which have space stations. Cities are numbered consecutively and each has a road of  length 1 km connecting it to the next city. It is not a circular route, so the first city doesn't connect with the last city. Determine the maximum distance from any city to it's nearest space station.